Provide new "load_email_data" after-event for new Gmail.
The load_email_data event intercepts completed /i/fd-requests
and parses the response-data to provide useful email/thread information.

This data can later be cached and used as a data-store for a new
api.get.email_data() which does not have to depend on forged
XHR-requests to provide rich email meta-data.